feat: :sparkles: Sub SurrealDB for Postgres + Shuttle
This PR contains the changes necessary to remove SurrealDB from the project and replace it with PostgreSQL as a DB used in the hosting platform shuttle.

+++ utils/src/lib.rs
@@ -1,8 +1,8 @@
 use serde::{Deserialize, Serialize};
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
 pub struct Entry {
     pub name: String,
-    pub score: usize,
+    pub score: i64,
 }
 
 #[cfg(test)]
